requirements.txt
nbconvert==7.16.4
pelican==4.10.1
markdown==3.7
gitpython==3.1.43